+++ pkg/util/validation/limits.go
@@ -21,33 +21,30 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/grafana/mimir/pkg/ingester/activeseries"
-	"github.com/grafana/mimir/pkg/mimirpb"
 	"github.com/grafana/mimir/pkg/storage/tsdb/block"
-	"github.com/grafana/mimir/pkg/util/ephemeral"
 )
 
 const (
-	MaxSeriesPerMetricFlag        = "ingester.max-global-series-per-metric"
-	MaxMetadataPerMetricFlag      = "ingester.max-global-metadata-per-metric"
-	MaxSeriesPerUserFlag          = "ingester.max-global-series-per-user"
-	MaxEphemeralSeriesPerUserFlag = "ingester.max-ephemeral-series-per-user"
-	MaxMetadataPerUserFlag        = "ingester.max-global-metadata-per-user"
-	MaxChunksPerQueryFlag         = "querier.max-fetched-chunks-per-query"
-	MaxChunkBytesPerQueryFlag     = "querier.max-fetched-chunk-bytes-per-query"
-	MaxSeriesPerQueryFlag         = "querier.max-fetched-series-per-query"
-	maxLabelNamesPerSeriesFlag    = "validation.max-label-names-per-series"
-	maxLabelNameLengthFlag        = "validation.max-length-label-name"
-	maxLabelValueLengthFlag       = "validation.max-length-label-value"
-	maxMetadataLengthFlag         = "validation.max-metadata-length"
-	creationGracePeriodFlag       = "validation.create-grace-period"
-	maxQueryLengthFlag            = "store.max-query-length"
-	maxPartialQueryLengthFlag     = "querier.max-partial-query-length"
-	maxTotalQueryLengthFlag       = "query-frontend.max-total-query-length"
-	requestRateFlag               = "distributor.request-rate-limit"
-	requestBurstSizeFlag          = "distributor.request-burst-size"
-	ingestionRateFlag             = "distributor.ingestion-rate-limit"
-	ingestionBurstSizeFlag        = "distributor.ingestion-burst-size"
-	HATrackerMaxClustersFlag      = "distributor.ha-tracker.max-clusters"
+	MaxSeriesPerMetricFlag     = "ingester.max-global-series-per-metric"
+	MaxMetadataPerMetricFlag   = "ingester.max-global-metadata-per-metric"
+	MaxSeriesPerUserFlag       = "ingester.max-global-series-per-user"
+	MaxMetadataPerUserFlag     = "ingester.max-global-metadata-per-user"
+	MaxChunksPerQueryFlag      = "querier.max-fetched-chunks-per-query"
+	MaxChunkBytesPerQueryFlag  = "querier.max-fetched-chunk-bytes-per-query"
+	MaxSeriesPerQueryFlag      = "querier.max-fetched-series-per-query"
+	maxLabelNamesPerSeriesFlag = "validation.max-label-names-per-series"
+	maxLabelNameLengthFlag     = "validation.max-length-label-name"
+	maxLabelValueLengthFlag    = "validation.max-length-label-value"
+	maxMetadataLengthFlag      = "validation.max-metadata-length"
+	creationGracePeriodFlag    = "validation.create-grace-period"
+	maxQueryLengthFlag         = "store.max-query-length"
+	maxPartialQueryLengthFlag  = "querier.max-partial-query-length"
+	maxTotalQueryLengthFlag    = "query-frontend.max-total-query-length"
+	requestRateFlag            = "distributor.request-rate-limit"
+	requestBurstSizeFlag       = "distributor.request-burst-size"
+	ingestionRateFlag          = "distributor.ingestion-rate-limit"
+	ingestionBurstSizeFlag     = "distributor.ingestion-burst-size"
+	HATrackerMaxClustersFlag   = "distributor.ha-tracker.max-clusters"
 
 	// MinCompactorPartialBlockDeletionDelay is the minimum partial blocks deletion delay that can be configured in Mimir.
 	MinCompactorPartialBlockDeletionDelay = 4 * time.Hour
@@ -94,8 +91,6 @@ type Limits struct {
 	// Series
 	MaxGlobalSeriesPerUser   int `yaml:"max_global_series_per_user" json:"max_global_series_per_user"`
 	MaxGlobalSeriesPerMetric int `yaml:"max_global_series_per_metric" json:"max_global_series_per_metric"`
-	// Ephemeral series
-	MaxEphemeralSeriesPerUser int `yaml:"max_ephemeral_series_per_user" json:"max_ephemeral_series_per_user" category:"experimental"`
 	// Metadata
 	MaxGlobalMetricsWithMetadataPerUser int `yaml:"max_global_metadata_per_user" json:"max_global_metadata_per_user"`
 	MaxGlobalMetadataPerMetric          int `yaml:"max_global_metadata_per_metric" json:"max_global_metadata_per_metric"`
@@ -177,8 +172,6 @@ type Limits struct {
 	ForwardingEndpoint      string          `yaml:"forwarding_endpoint" json:"forwarding_endpoint" doc:"nocli|description=Remote-write endpoint where metrics specified in forwarding_rules are forwarded to. If set, takes precedence over endpoints specified in forwarding rules."`
 	ForwardingDropOlderThan model.Duration  `yaml:"forwarding_drop_older_than" json:"forwarding_drop_older_than" doc:"nocli|description=If set, forwarding drops samples that are older than this duration. If unset or 0, no samples get dropped."`
 	ForwardingRules         ForwardingRules `yaml:"forwarding_rules" json:"forwarding_rules" doc:"nocli|description=Rules based on which the Distributor decides whether a metric should be forwarded to an alternative remote_write API endpoint."`
-
-	EphemeralSeriesMatchers ephemeral.LabelMatchers `yaml:"ephemeral_series_matchers" json:"ephemeral_series_matchers" category:"experimental"`
 }
 
 // RegisterFlags adds the flags required to config this to the given FlagSet
@@ -203,7 +196,6 @@ func (l *Limits) RegisterFlags(f *flag.FlagSet) {
 
 	f.IntVar(&l.MaxGlobalSeriesPerUser, MaxSeriesPerUserFlag, 150000, "The maximum number of in-memory series per tenant, across the cluster before replication. 0 to disable.")
 	f.IntVar(&l.MaxGlobalSeriesPerMetric, MaxSeriesPerMetricFlag, 0, "The maximum number of in-memory series per metric name, across the cluster before replication. 0 to disable.")
-	f.IntVar(&l.MaxEphemeralSeriesPerUser, MaxEphemeralSeriesPerUserFlag, 0, "The maximum number of in-memory ephemeral series per tenant, across the cluster before replication. 0 to disable ephemeral storage.")
 
 	f.IntVar(&l.MaxGlobalMetricsWithMetadataPerUser, MaxMetadataPerUserFlag, 0, "The maximum number of in-memory metrics with metadata per tenant, across the cluster. 0 to disable.")
 	f.IntVar(&l.MaxGlobalMetadataPerMetric, MaxMetadataPerMetricFlag, 0, "The maximum number of metadata per metric, across the cluster. 0 to disable.")
@@ -270,8 +262,6 @@ func (l *Limits) RegisterFlags(f *flag.FlagSet) {
 	f.IntVar(&l.AlertmanagerMaxDispatcherAggregationGroups, "alertmanager.max-dispatcher-aggregation-groups", 0, "Maximum number of aggregation groups in Alertmanager's dispatcher that a tenant can have. Each active aggregation group uses single goroutine. When the limit is reached, dispatcher will not dispatch alerts that belong to additional aggregation groups, but existing groups will keep working properly. 0 = no limit.")
 	f.IntVar(&l.AlertmanagerMaxAlertsCount, "alertmanager.max-alerts-count", 0, "Maximum number of alerts that a single tenant can have. Inserting more alerts will fail with a log message and metric increment. 0 = no limit.")
 	f.IntVar(&l.AlertmanagerMaxAlertsSizeBytes, "alertmanager.max-alerts-size-bytes", 0, "Maximum total size of alerts that a single tenant can have, alert size is the sum of the bytes of its labels, annotations and generatorURL. Inserting more alerts will fail with a log message and metric increment. 0 = no limit.")
-
-	f.Var(&l.EphemeralSeriesMatchers, "distributor.ephemeral-series-matchers", fmt.Sprintf("Lists of series matchers prefixed by the source. The source must be one of %s. If an incoming sample matches at least one of the matchers with its source it gets marked as ephemeral. The format of the value looks like: %s:{namespace=\"dev\"};%s:{host=\"server1\",namespace=\"prod\"}", strings.Join(ephemeral.ValidSourceStrings, ", "), ephemeral.API, ephemeral.RULE))
 }
 
 // UnmarshalYAML implements the yaml.Unmarshaler interface.
@@ -463,11 +453,6 @@ func (o *Overrides) MaxGlobalSeriesPerUser(userID string) int {
 // MaxGlobalSeriesPerMetric returns the maximum number of series allowed per metric across the cluster.
 func (o *Overrides) MaxGlobalSeriesPerMetric(userID string) int {
 	return o.getOverridesForUser(userID).MaxGlobalSeriesPerMetric
-}
-
-// MaxEphemeralSeriesPerUser returns the maximum number of ephemeral series a user is allowed to store across the cluster.
-func (o *Overrides) MaxEphemeralSeriesPerUser(userID string) int {
-	return o.getOverridesForUser(userID).MaxEphemeralSeriesPerUser
 }
 
 func (o *Overrides) MaxChunksPerQuery(userID string) int {
@@ -792,15 +777,6 @@ func (o *Overrides) ForwardingRules(user string) ForwardingRules {
 
 func (o *Overrides) ForwardingEndpoint(user string) string {
 	return o.getOverridesForUser(user).ForwardingEndpoint
-}
-
-func (o *Overrides) EphemeralChecker(user string, source mimirpb.WriteRequest_SourceEnum) ephemeral.SeriesChecker {
-	m := o.getOverridesForUser(user).EphemeralSeriesMatchers.ForSource(source)
-	if m.HasMatchers() {
-		return &m
-	}
-
-	return nil
 }
 
 func (o *Overrides) ForwardingDropOlderThan(user string) time.Duration {